+++ src/mexlist.c
@@ -335,7 +335,7 @@ static void _finished_cb(MGEFF_ANIMATION animation)
 }
 
 
-static void _call_draw(MGEFF_ANIMATION animation, void* target, int id, void* value)
+static void _call_draw(MGEFF_ANIMATION animation, void* target, intptr_t id, void* value)
 {
     UpdateWindow((HWND)target, TRUE);
 }
